Use annotations from future for cleaner type hint. Moved qe_time_to_sec to utils. Renamed BaseStdoutParser.parse_stdout_base to parse for compatibility with parent abstract class.
