+++ src/qe_tools/converters/pymatgen.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from __future__ import annotations
+
 from importlib.util import find_spec
 
 _has_pmg = bool(find_spec('pymatgen'))
