+++ src/qe_tools/inputs/base.py
@@ -5,6 +5,8 @@
 Tools for parsing QE PW input files.
 """
 
+from __future__ import annotations
+
 import re
 from typing import Dict, Iterable, List, Optional
