+++ src/qe_tools/inputs/pw.py
@@ -3,6 +3,8 @@
 Tools for parsing QE PW input files
 """
 
+from __future__ import annotations
+
 import re
 
 from qe_tools.exceptions import ParsingError
