+++ src/qe_tools/outputs/base.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 """Base parser for the outputs of Quantum ESPRESSO."""
 
-from typing import Union
+from __future__ import annotations
+
 import abc
 
 
@@ -10,7 +11,7 @@ class BaseOutput(abc.ABC):
     Abstract class for the outputs of Quantum ESPRESSO.
     """
 
-    def __init__(self, outputs: Union[dict, None] = None, executable: Union[str, None] = None):
+    def __init__(self, outputs: dict | None = None, executable: str | None = None):
         self.outputs = outputs
         self.executable = executable
